Move the red-black tree code to a separate package
